requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
django-cors-headers==2.1.0
django-countries==4.6.1
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
PyJWT==1.5.2
python-decouple==3.0
pytz==2017.2
requests==2.18.2
uritemplate==3.0.0
urllib3==1.22
